Increased attempts to 128 and made the number of random instructions increase linearly as the age increases.

+++ FireStarter/FireStarterEvolve.h
@@ -13,8 +13,8 @@ private:
 
 public:
 	bool EvolveSeeds(const FireStarterSettings& settings, bool sync = false);
-	bool EvolveState(const FireStarterState& state, size_t generation, bool sync = false);
-	bool EvolveStates(const FireStarterState& bestState, const std::vector<FireStarterState>& allStates, size_t generation, bool sync = false);
+	bool EvolveState(const FireStarterState& state, unsigned long long generation, bool sync = false);
+	bool EvolveStates(const FireStarterState& bestState, const std::vector<FireStarterState>& allStates, unsigned long long generation, bool sync = false);
 	bool GenerateOptimize(const FireStarterState& initState, bool sync = false);
 	FireStarterEvolve(FireStarterManager* manager, size_t index = 0);
 	~FireStarterEvolve(void);
